refactor: Change float to double in Distance and Position

+++ Includes/CubbyTower/Components/Distance.hpp
@@ -16,7 +16,7 @@ namespace CubbyTower
 //!
 struct Distance
 {
-    float s;
+    double s;
 };
 }  // namespace CubbyTower
